requirements.txt
torch>=2.4.0
torchvision>=0.19.0
diffusers>=0.37.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0
gradio>=5.0.0
mediapipe>=0.10.0
opencv-python-headless>=4.9.0
insightface>=0.7.0
onnxruntime-gpu>=1.18.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
sentencepiece>=0.2.0
protobuf>=4.25.0
piq>=0.8.0
lpips>=0.1.4
scikit-image>=0.22.0
